feat: Time labels have proper click events
Click events for time labels read as a switch statement and change variables this.miltime and this.meridiem.
After values are changed in switch statement, ChangeTimeLabels is called with the values passed as arguments
Currently ChangeTimeLabels outputs whether military time is true or not.
TODO: Change styles and click functions based on variable values